Game enchancements
- Added a speedMultiplier to the Ball script.
- Modified Ball object Collision Detection property to Continuous.
- Modified Bounce Threshold to 0.01 in Project Settings.

+++ Assets/Scripts/Paddle.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class Paddle : MonoBehaviour
 {
     public new Rigidbody2D rigidbody2D;
-    public float id;
+    public int id;
     public float moveSpeed = 2f;
 
     public void Update()
